feat: Implement interactive configuration wizard and enhance logging output
