+++ src/DevOpsMigrationPlatform.CLI.Migration/Commands/LogsCommand.cs
@@ -5,6 +5,7 @@ using DevOpsMigrationPlatform.CLI.Migration.Commands;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
+using Spectre.Console;
 using Spectre.Console.Cli;
 using System.ComponentModel;
 
@@ -42,9 +43,11 @@ public sealed class LogsCommand : CommandBase<LogsCommand.Settings>
             services.AddSingleton<ILogsClient>(sp => sp.GetRequiredService<ControlPlaneClient>());
         });
 
+        var console = GetRequiredService<IAnsiConsole>();
+
         try
         {
-            return await RunCoreAsync(settings);
+            return await RunCoreAsync(settings, console);
         }
         catch (OperationCanceledException)
         {
@@ -52,12 +55,12 @@ public sealed class LogsCommand : CommandBase<LogsCommand.Settings>
         }
         catch (HttpRequestException ex)
         {
-            Console.Error.WriteLine($"Error: {ex.Message}");
+            console.MarkupLine($"[red]Error: {Markup.Escape(ex.Message)}[/]");
             return 1;
         }
     }
 
-    private async Task<int> RunCoreAsync(Settings settings)
+    private async Task<int> RunCoreAsync(Settings settings, IAnsiConsole console)
     {
         var client = GetRequiredService<ILogsClient>();
         using var cts = new CancellationTokenSource();
@@ -67,12 +70,12 @@ public sealed class LogsCommand : CommandBase<LogsCommand.Settings>
         {
             var events = await client.GetLogsAsync(settings.JobId, cts.Token);
             foreach (var evt in events)
-                Console.WriteLine(JsonSerializer.Serialize(evt, _jsonOptions));
+                console.WriteLine(JsonSerializer.Serialize(evt, _jsonOptions));
             return 0;
         }
 
         await foreach (var evt in client.FollowLogsAsync(settings.JobId, cts.Token))
-            Console.WriteLine(JsonSerializer.Serialize(evt, _jsonOptions));
+            console.WriteLine(JsonSerializer.Serialize(evt, _jsonOptions));
 
         return 0;
     }
